standardize email pattern
