+++ pkg/detectors/satismeterprojectkey/satismeterprojectkey.go
@@ -4,9 +4,10 @@ import (
 	"context"
 	b64 "encoding/base64"
 	"fmt"
-	regexp "github.com/wasilibs/go-re2"
 	"net/http"
 	"strings"
+
+	regexp "github.com/wasilibs/go-re2"
 
 	"github.com/trufflesecurity/trufflehog/v3/pkg/common"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
@@ -25,7 +26,7 @@ var (
 
 	// Make sure that your group is surrounded in boundary characters such as below to reduce false positives.
 	keyPat   = regexp.MustCompile(detectors.PrefixRegex([]string{"satismeter"}) + `\b([a-zA-Z0-9]{24})\b`)
-	emailPat = regexp.MustCompile(detectors.PrefixRegex([]string{"satismeter"}) + `\b([a-zA-Z0-9]{4,20}@[a-zA-Z0-9]{2,12}.[a-zA-Z0-9]{2,12})\b`)
+	emailPat = regexp.MustCompile(detectors.PrefixRegex([]string{"satismeter"}) + common.EmailPattern)
 	passPat  = regexp.MustCompile(detectors.PrefixRegex([]string{"satismeter"}) + `\b([a-zA-Z0-9!=@#$%^]{6,32})`)
 )
 
@@ -39,40 +40,34 @@ func (s Scanner) Keywords() []string {
 func (s Scanner) FromData(ctx context.Context, verify bool, data []byte) (results []detectors.Result, err error) {
 	dataStr := string(data)
 
-	matches := keyPat.FindAllStringSubmatch(dataStr, -1)
-	emailmatches := emailPat.FindAllStringSubmatch(dataStr, -1)
-	passmatches := passPat.FindAllStringSubmatch(dataStr, -1)
+	uniqueEmailMatches, uniqueKeyMatches, uniquePassMatches := make(map[string]struct{}), make(map[string]struct{}), make(map[string]struct{})
+	for _, match := range emailPat.FindAllStringSubmatch(dataStr, -1) {
+		uniqueEmailMatches[strings.TrimSpace(match[1])] = struct{}{}
+	}
 
-	for _, match := range matches {
-		if len(match) != 2 {
-			continue
-		}
-		resMatch := strings.TrimSpace(match[1])
+	for _, match := range keyPat.FindAllStringSubmatch(dataStr, -1) {
+		uniqueKeyMatches[strings.TrimSpace(match[1])] = struct{}{}
+	}
 
-		for _, emailmatch := range emailmatches {
-			if len(emailmatch) != 2 {
-				continue
-			}
-			resEmailMatch := strings.TrimSpace(emailmatch[1])
+	for _, match := range passPat.FindAllStringSubmatch(dataStr, -1) {
+		uniquePassMatches[strings.TrimSpace(match[1])] = struct{}{}
+	}
 
-			for _, passmatch := range passmatches {
-				if len(passmatch) != 2 {
-					continue
-				}
-				resPassMatch := strings.TrimSpace(passmatch[1])
-
+	for keyMatch := range uniqueKeyMatches {
+		for emailMatch := range uniqueEmailMatches {
+			for passMatch := range uniquePassMatches {
 				s1 := detectors.Result{
 					DetectorType: detectorspb.DetectorType_SatismeterProjectkey,
-					Raw:          []byte(resMatch),
-					RawV2:        []byte(resMatch + resPassMatch),
+					Raw:          []byte(keyMatch),
+					RawV2:        []byte(keyMatch + passMatch),
 				}
 
 				if verify {
 
-					data := fmt.Sprintf("%s:%s", resEmailMatch, resPassMatch)
+					data := fmt.Sprintf("%s:%s", emailMatch, passMatch)
 					sEnc := b64.StdEncoding.EncodeToString([]byte(data))
 
-					req, err := http.NewRequestWithContext(ctx, "GET", "https://app.satismeter.com/api/users?project="+resMatch, nil)
+					req, err := http.NewRequestWithContext(ctx, "GET", "https://app.satismeter.com/api/users?project="+keyMatch, nil)
 					if err != nil {
 						continue
 					}
